requirements.txt
bzip2=1.0.8
ca-certificates=2025.7.15
certifi=2022.12.7
charset-normalizer=2.1.1
cmake=3.25.0
einops=0.8.1
einx=0.3.0
filelock=3.13.1
frozendict=2.4.6
fsspec=2025.7.0
ftfy=6.3.1
hf-xet=1.1.7
huggingface-hub=0.34.4
idna=3.4
jinja2=3.1.4
joblib=1.5.1
ld_impl_linux-64=2.40
libffi=3.3
libgcc-ng=11.2.0
libgomp=11.2.0
libstdcxx-ng=11.2.0
libuuid=1.41.5
libxcb=1.17.0
lit=15.0.7
markupsafe=2.1.5
mpmath=1.3.0
ncurses=6.5
networkx=3.3
numpy=2.1.2
opencv-python=4.12.0.88
openssl=1.1.1w
packaging=25.0
pillow=11.0.0
pip=25.1
pthread-stubs=0.3
python=3.10.0
pyyaml=6.0.2
readline=8.3
regex=2025.7.34
requests=2.28.1
safetensors=0.6.2
scikit-learn=1.7.1
scipy=1.15.3
setuptools=78.1.1
sqlite=3.50.2
sympy=1.13.3
threadpoolctl=3.6.0
timm=1.0.19
tk=8.6.15
torch=2.0.1+cu117
torchaudio=2.0.2+cu117
torchvision=0.15.2+cu117
tqdm=4.67.1
triton=2.0.0
typing-extensions=4.12.2
tzdata=2025b
urllib3=1.26.13
wcwidth=0.2.13
wheel=0.45.1
xorg-libx11=1.8.12
xorg-libxau=1.0.12
xorg-libxdmcp=1.1.5
xorg-xorgproto=2024.1
xz=5.6.4
zlib=1.2.13
